Abstracted mode switching logic
